.archive_wp_yemek/server/requirements.txt
fastapi
uvicorn[standard]
httpx==0.27.*
beautifulsoup4
python-dotenv
pydantic
python-dateutil
pillow
pytesseract
openai